requirements.txt
openenv-core==0.2.3
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.11.7
openai>=2.30.0